services/agrometeo-simulator-py/requirements.txt
paho-mqtt==1.6.1
perlin-noise==1.12
